[EDELIVERY-4987] implement some Sonar suggestions

+++ Domibus-MSH/src/main/java/eu/domibus/web/rest/PropertyResource.java
@@ -23,8 +23,8 @@ import java.util.stream.Collectors;
 /**
  * @author Ion Perpegel
  * @since 4.1.1
- *
- * Resource responsible with getting the domibus properties that can be changed at runtime, getting and setting their values through REST Api
+ * <p>
+ * Resource responsible for getting the domibus properties that can be changed at runtime, getting and setting their values through REST Api
  */
 @RestController
 @RequestMapping(value = "/rest/configuration")
@@ -49,7 +49,7 @@ public class PropertyResource {
         List<DomibusProperty> items = domibusPropertyService.getProperties(request.getName());
         response.setCount(items.size());
         items = items.stream()
-                .skip(request.getPage() * request.getPageSize())
+                .skip((long) request.getPage() * request.getPageSize())
                 .limit(request.getPageSize())
                 .collect(Collectors.toList());
         response.setItems(domainConverter.convert(items, DomibusPropertyRO.class));
